Simplify publish.js
